% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eodf_geturl_dl.R
\name{eodf_geturl_dl}
\alias{eodf_geturl_dl}
\title{Get URL and download}
\usage{
eodf_geturl_dl(dates, lag_time, roi, roi_name, collection, variables,
  parallel = TRUE, verbose = FALSE)
}
\arguments{
\item{verbose}{}
}
\description{

}
