% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_var_on_buffers.R
\name{extract_var_on_buffers}
\alias{extract_var_on_buffers}
\title{Extract variables given a vector of buffer sizes}
\usage{
extract_var_on_buffers(rastBrick, sf_points, buffer_sizes,
  fun_summarize = "mean", is_timeseries = FALSE,
  date_sampling = NULL, na_max_perc = 0, code_indice = NULL,
  verbose = FALSE)
}
\arguments{
\item{rastBrick}{list of raster bricks (1 to n)}

\item{sf_points}{list of sampling sf_points (1 to n)}

\item{buffer_sizes}{vector of buffers (1 to z)}

\item{fun_summarize}{summarizing function}

\item{date_sampling}{list of sampling dates (1 to n)}

\item{na_max_perc}{integer. if, in a given buffer, there are more than na_max_perc cells that are NA, the overall value for this buffer will be NA. default is 0 (i.e. as long as there is one cell with a value, the overall value for the buffer will be computed)}

\item{code_indice}{character vector. code indice (only for timeseries)}

\item{verbose}{}
}
\description{

}
\details{
Argument RastBrick is either :
- a Raster object
- a rasterBrick object
}
