% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_df_points.R
\name{prepare_df_points}
\alias{prepare_df_points}
\title{Prepare the input data.frame of sampling points for the rest of the workflow}
\usage{
prepare_df_points(df_points_metadata)
}
\arguments{
\item{df_points_metadata}{data.frame. Contains the input sampling points. Expected format is given in the details}

\item{lag_time}{integer. lagged time (in days) for which data will be extracted from the time series collections}
}
\description{
Prepare the input data.frame of sampling points for the rest of the workflow
}
\details{
Expected format for input data.frame \code{df_points_metadata} is the following :
\describe{
\item{id}{character string. unique identifier of the sampling point}
\item{date}{sampling date (character string) }
\item{longitude}{longitude}
\item{latitude}{latitude}
}
}
