% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mixed_products.R
\name{prepare_mixed_products}
\alias{prepare_mixed_products}
\title{Prepare mixed products (e.g. MODIS TERRA and AQUA)}
\usage{
prepare_mixed_products(rasts_terra, rasts_aqua, fun_summarize)
}
\arguments{
\item{fun_summarize}{}
}
\description{

}
