% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_modis_mixed_products.R
\name{prepare_modis_mixed_products}
\alias{prepare_modis_mixed_products}
\title{Prepare MODIS mixed Terra and Aqua products}
\usage{
prepare_modis_mixed_products(rasts_terra, rasts_aqua, fun_summarize)
}
\arguments{
\item{fun_summarize}{}
}
\description{

}
