% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_roi.R
\name{prepare_roi}
\alias{prepare_roi}
\title{Prepare the roi for the rest of the workflow}
\usage{
prepare_roi(df_points_metadata, buffer_sizes)
}
\arguments{
\item{df_points_metadata}{data.frame. Contains the input sampling points. Expected format is given in the details}

\item{buffer_sizes}{vector of buffer sizes}
}
\description{
Prepare the roi for the rest of the workflow
}
