% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_s2_indices.R
\name{prepare_s2_indices}
\alias{prepare_s2_indices}
\title{Create valid spectral indices from S2 data averaged over a given time period.}
\usage{
prepare_s2_indices(th_list_of_url, variables, spectral_indices)
}
\arguments{
\item{th_list_of_url}{data.frame, output of shub4r::shr_get_url()}

\item{variables}{character vector. vector of bands (e.g. c("BO4","B08"))}

\item{spectral_indices}{character vector. vector of spectral indices to compute (e.g. c("ndvi","mndvi"))}

\item{verbose}{verbose}
}
\value{
a rasterBrick with the spectral indices (1 band / indice).
}
\description{

}
\details{
param th_list_of_url is the output of shub4r::shr_get_url. It must contain the band "9_SCENE_CLASSIFICATION"
}
