% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_topography_indices.R
\name{prepare_topography_indices}
\alias{prepare_topography_indices}
\title{Prepare the topography indices from the DEM}
\usage{
prepare_topography_indices(path_to_input_dem_files, roi,
  path_to_grassApplications_folder)
}
\description{
Create the topography datasets from the DEM with the rgrass7 package : slope, aspect, accumulation, tci, twi
}
\details{
param path_to_grassApplications_folder : Path to the GRASS application folder, for the setup of the rgrass7 package. Can be retrieved in the terminal with grass74 --config path . More info on the use of rgrass7 at https://grasswiki.osgeo.org/wiki/R_statistics/rgrass7
}
